feat: 다른 유저의 즐겨찾기 삭제 방지 기능 추가

+++ src/main/java/nextstep/subway/favorite/application/FavoriteService.java
@@ -1,7 +1,6 @@
 package nextstep.subway.favorite.application;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 import org.springframework.stereotype.Service;
@@ -12,12 +11,12 @@ import nextstep.subway.favorite.domain.Favorite;
 import nextstep.subway.favorite.domain.FavoriteRepository;
 import nextstep.subway.favorite.dto.FavoriteRequest;
 import nextstep.subway.favorite.dto.FavoriteResponse;
+import nextstep.subway.favorite.exception.FavoriteNotFoundException;
+import nextstep.subway.favorite.exception.NotMineFavoriteException;
 import nextstep.subway.member.application.MemberService;
 import nextstep.subway.member.domain.Member;
-import nextstep.subway.member.domain.MemberRepository;
 import nextstep.subway.station.application.StationService;
 import nextstep.subway.station.domain.Station;
-import nextstep.subway.station.domain.StationRepository;
 
 @Service
 @Transactional
@@ -51,8 +50,15 @@ public class FavoriteService {
             .collect(Collectors.toList());
     }
 
-    public void deleteFavorite(Long id) {
-        Favorite favorite = favoriteRepository.findById(id).orElseThrow(NoSuchElementException::new);
+    private Favorite findById(Long id) {
+        return favoriteRepository.findById(id).orElseThrow(FavoriteNotFoundException::new);
+    }
+
+    public void deleteFavorite(LoginMember loginMember, Long id) {
+        Favorite favorite = findById(id);
+        if (!favorite.isOwner(loginMember.getId())) {
+            throw new NotMineFavoriteException();
+        }
         favoriteRepository.delete(favorite);
     }
 }
